# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rifa API Docs",
        "description": "API para gerenciamento de rifas",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "API Licence",
            "url": "https://github.com/felipesilva15/rifa-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ticket": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "List all tickets",
                "operationId": "0c4907ec7353d133ef22fb72c50f4e66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Ticket ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "participant_id",
                        "in": "query",
                        "description": "Participant ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "raffle_id",
                        "in": "query",
                        "description": "Raffle ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Ticket number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "payment_date",
                        "in": "query",
                        "description": "Ticket payment date",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Ticket value",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Registers an ticket",
                "operationId": "c3e39f1002b325b5b28c2335f69843d8",
                "requestBody": {
                    "description": "Data for creating a new ticket",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ticket"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered ticket data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ticket/{id}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "List an ticket by ID",
                "operationId": "6acf0d4d743f82777a7239b8ada53ecb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Update an ticket",
                "operationId": "554eae33f31f3bf26351e22196e59d10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for update ticket",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ticket"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated ticket data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Deletes an ticket",
                "operationId": "7988715938e134d642195ed746d5cb24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Response"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error ocorried."
                    },
                    "route": {
                        "type": "string",
                        "example": "api/error"
                    }
                },
                "type": "object"
            },
...
            "Ticket": {
                "required": [
                    "raffle_id",
                    "participant_id",
                    "number"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "raffle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "participant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "number": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "payment_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-01"
                    },
                    "value": {
                        "type": "number",
                        "format": "double",
                        "maximum": 999.99,
                        "minimum": 0.01,
                        "example": 20
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T03:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30T03:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}